Perform mining through a mining pool (#32)
This way, the mining difficulty can be set to a number high enough that
makes mining transactions possible.

This is a workaround to the preemptive block template generation that
often privileges mining empty blocks, even if transactions are available
in the mempool, effectively making it impossible in certain
circumstances to mine any transaction. This workaround may be removed
(in favor of using a standalone miner) once the preemptive block
generation can be disabled.