When testing mints with transferOwnershipTo before a hard fork, make sure that the mint expires before the hard fork
If a mint with transferOwnershipTo is produced close to the hard fork
block, it may be accepted into the mempool even if may not be mined
right away. The simulation expects the mint to be rejected, so set the
expiration before the hard fork to make sure it always gets rejected.