Workaround for situations in which nodes may not realize they're out-of-sync
In some circumstances, nodes may be out-of-sync without realizing it.
When the happens, they will not make any attempt at fetching missing
blocks, and calling `waitForConvergence` will fail with a timeout.

This commit implements a workaround for that problem: once a timeout is
detected, a new block is mined, in an attempt to get the nodes aware
that the chain has changed and get them unstuck.